requirements.txt
openai
transformers
torch
xformers
accelerate
black
pytype
GitPython